# api/openapi.yaml
openapi: 3.0.3
info:
  title: RefMD API
  description: Real-time collaborative markdown editor API with CRDT support
  version: 1.1.0
  contact:
    name: RefMD Team
    email: support@refmd.io
servers:
  - url: http://localhost:8888/api
    description: Development server
  - url: https://api.refmd.io/api
    description: Production server
paths:
...
  /documents:
    get:
      tags:
        - Documents
      summary: List user documents
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, title]
            default: updated_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Documents
      summary: Create a new document
      operationId: createDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      operationId: getDocument
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      responses:
        '200':
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Documents
      summary: Update document
      operationId: updateDocument
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for updating shared documents
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Documents
      summary: Delete document
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/content:
    get:
      tags:
        - Documents
      summary: Get document content
      description: Returns the current document content as markdown text
      operationId: getDocumentContent
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      responses:
        '200':
          description: Document content retrieved successfully
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/state:
    get:
      tags:
        - Documents
      summary: Get document CRDT state
      description: Returns the current CRDT state vector for synchronization
      operationId: getDocumentState
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      responses:
        '200':
          description: State vector retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/updates:
    post:
      tags:
        - Documents
      summary: Get document updates since timestamp
      description: Returns incremental CRDT updates since a given timestamp
      operationId: getDocumentUpdates
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                since:
                  type: integer
                  format: int64
                  description: Unix timestamp in milliseconds
              required:
                - since
      responses:
        '200':
          description: Updates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updates:
                    type: array
                    items:
                      type: string
                      format: base64
                  timestamp:
                    type: integer
                    format: int64
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download document with attachments
      description: Downloads the document and all its attachments as a ZIP file
      operationId: downloadDocument
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          description: Share token for accessing shared documents
          schema:
            type: string
      responses:
        '200':
          description: ZIP file containing document and attachments
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/backlinks:
    get:
      tags:
        - Documents
      summary: Get backlinks for a document
      description: Returns all documents that link to this document
      operationId: getDocumentBacklinks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Backlinks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacklinksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/links:
    get:
      tags:
        - Documents
      summary: Get outgoing links from a document
      description: Returns all documents that this document links to
      operationId: getDocumentLinks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Outgoing links retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutgoingLinksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateDocumentRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        content:
          type: string
          default: ""
        type:
          type: string
          enum: [document, folder, scrap]
          default: document
        parent_id:
          type: string
          format: uuid
          nullable: true
      required:
        - title
...
    DocumentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer

    # ===== File =====
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...